Store prices as integer ticks plus per-instrument tick size (#155)
Every canonical price column becomes int64 ticks; the quote-currency price is
ticks * tick_size, where tick_size is the instrument's minimum price increment
(PipelineConfig.tick_size, default 0.01). Loaders quantise raw prices to ticks
on load; the plots and round-trip writers convert back for display. Storing the
exact integer removes the float rounding that made small-tick and 0-1
instruments show crossed levels that were not real, and the depth engine now
bins and compares levels on exact integers instead of multiplying and rounding
each event.

- config: add tick_size; reframe price_divisor as the raw-feed encoding scale
  and price_decimals as display precision; price_multiplier is now the integer
  inverse of tick_size.
- _utils: price_to_ticks / ticks_to_price / tick_multiplier (the exact integer
  path reproduces the engine's former internal integers).
- loaders (bitstamp, lobster, depth_l2): parse to ticks, writers invert;
  LOBSTER's price_divisor is now just the raw encoding, generalised into the one
  tick rule.
- depth engine: drop the multiply-and-round path; best prices stay int64.
- synth emits its internal ticks; toy datasets use tick_size 1.0 so the
  teaching prices stay readable.
- analytics read ticks: scale-free outputs (bps depth, order-book imbalance,
  VPIN) are unchanged; price-valued ones (trade_impacts VWAP, Kyle's lambda) are
  now in tick units, documented.
- plots convert once via the new public, idempotent display_result, used by
  build_gallery_model and the low-level tutorial/notebook examples.
- schema version 2.0 -> 3.0; tick size stored in Parquet key-value metadata (a
  JSON map keyed by instrument, ready for per-(venue, symbol) ticks in #147) and
  surfaced on load_data frames' attrs; older float-price files still read.
- docs: schema.md price policy, every price row, and decision #3; CHANGELOG;
  tutorial LOBSTER-writer fix; custom-components and protocol docstrings. Ignore
  B018 for tutorial jupytext cells (bare display expressions, like the existing
  E402 import exemption).

The reconstructed book is unchanged, only re-expressed: verified element-wise
against the pre-tick float pipeline, where ticks * tick_size reproduces the old
events / depth / depth_summary prices exactly, with every volume and row
identical.

+++ tests/test_datasets.py
@@ -80,7 +80,7 @@ class TestLayout:
         assert events["event_id"].dtype == "int64"
         assert str(events["timestamp"].dtype) == "datetime64[ns, UTC]"
         assert str(events["exchange_timestamp"].dtype) == "datetime64[ns, UTC]"
-        assert events["price"].dtype == "float64"
+        assert events["price"].dtype == "int64"
         assert events["volume"].dtype == "float64"
         assert events["fill"].dtype == "float64"
         assert events["action"].dtype == "category"
